+++ CHANGELOG.md
@@ -14,6 +14,7 @@ v4.2
 - Associate motion data now handles orientation sensors as well so that orientation triads follow motions
 - When recording videos, if the user starts playing back a motion, recording is restarted to first animation frame.
 - Allow users to change video format from scripting shell (support gif, jpg, png with setVisualizerOption("video_format", "png")). 
+- Add Tools for Calibrating model based on IMU data and to solve Inverse Kinematics problem from IMU data.
 
 v4.1
 ====